fix: type vector bug with cor

+++ src/primary_objects.cpp
@@ -6,7 +6,7 @@
 
 using namespace std;
 
-int create_point(int x, int y, vector<float> cor, draws &structure_list) {
+int create_point(int x, int y, vector<double> cor, draws &structure_list) {
     ponto p;
     p.x = x;
     p.y = y;
@@ -15,7 +15,7 @@ int create_point(int x, int y, vector<float> cor, draws &structure_list) {
     return structure_list.lista_pontos.size() - 1;
 }
 
-int create_line(int x1, int y1, int x2, int y2, vector<float> cor1, vector<float> cor2, draws &structure_list) {
+int create_line(int x1, int y1, int x2, int y2, vector<double> cor1, vector<double> cor2, draws &structure_list) {
     ponto p1, p2;
 
     p1.x = x1;
